requirements.txt
python-telegram-bot==20.7
gspread==5.12.0
google-auth==2.23.4
fpdf2==2.7.6
python-dotenv==1.0.0
flask==3.0.0 